Move trips to App.jsx and import as prop
